Create about page instead of edit profile
