+++ app/src/main/java/com/bangkit/skutapplication/datastore/ViewModelFactory.kt
@@ -6,16 +6,14 @@ import com.bangkit.skutapplication.view.confirm.ConfirmViewModel
 import com.bangkit.skutapplication.view.login.LoginViewModel
 import com.bangkit.skutapplication.view.main.MainViewModel
 import com.bangkit.skutapplication.view.profile.ProfileViewModel
-import com.bangkit.skutapplication.view.profile.editprofile.EditProfileViewModel
+
 
 class ViewModelFactory(private val pref: UserPreference) : ViewModelProvider.NewInstanceFactory() {
 
     @Suppress("UNCHECKED_CAST")
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         return when {
-            modelClass.isAssignableFrom(EditProfileViewModel::class.java) -> {
-                EditProfileViewModel(pref) as T
-            }
+
             modelClass.isAssignableFrom(ProfileViewModel::class.java) -> {
                 ProfileViewModel(pref) as T
             }
