+++ app/src/main/java/com/bangkit/skutapplication/view/profile/ProfileViewModel.kt
@@ -18,4 +18,10 @@ class ProfileViewModel(private val pref: UserPreference): ViewModel() {
             pref.setFirstCalendar()
         }
     }
+
+    fun logout(){
+        viewModelScope.launch {
+            pref.logout()
+        }
+    }
 }
