+++ app/src/main/java/com/bangkit/skutapplication/view/profile/editprofile/AboutActivity.kt
@@ -1,0 +1,26 @@
+package com.bangkit.skutapplication.view.profile.editprofile
+
+import androidx.appcompat.app.AppCompatActivity
+import android.os.Bundle
+import android.view.MenuItem
+import com.bangkit.skutapplication.R
+import com.bangkit.skutapplication.databinding.ActivityEditProfileBinding
+
+
+class AboutActivity : AppCompatActivity() {
+    private lateinit var binding: ActivityEditProfileBinding
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_edit_profile)
+        binding = ActivityEditProfileBinding.inflate(layoutInflater)
+        setContentView(binding.root)
+        supportActionBar?.setDisplayHomeAsUpEnabled(true)
+    }
+    override fun onOptionsItemSelected(item: MenuItem): Boolean {
+        if (item.itemId == android.R.id.home) {
+            finish()
+        }
+        return super.onOptionsItemSelected(item)
+    }
+
+}
