+++ app/src/main/res/layout/fragment_profile.xml
@@ -79,9 +79,9 @@
         android:id="@+id/buttonEdit"
         android:layout_width="300dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="20dp"
+        android:layout_marginTop="10dp"
         android:backgroundTint="?android:textColor"
-        android:drawableEnd="@drawable/ic_baseline_edit_24"
+        android:drawableEnd="@drawable/ic_baseline_phone_android_24"
         android:text="ABOUT"
         android:textColor="@color/blue"
         app:layout_constraintEnd_toEndOf="@+id/buttonLanguage"
@@ -94,7 +94,7 @@
         android:id="@+id/buttonLanguage"
         android:layout_width="300dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="72dp"
+        android:layout_marginTop="30dp"
         android:backgroundTint="?android:textColor"
         android:drawableEnd="@drawable/ic_baseline_g_translate_24"
         android:text="@string/change_language"
@@ -110,7 +110,7 @@
         android:id="@+id/buttonLogout"
         android:layout_width="0dp"
         android:layout_height="wrap_content"
-        android:layout_marginTop="40dp"
+        android:layout_marginTop="70dp"
         android:drawableEnd="@drawable/ic_baseline_logout_24"
         android:text="@string/logout"
         android:textColor="@color/blue_200"
